+++ rope/base/oi/type_hinting/providers/interfaces.py
@@ -1,4 +1,4 @@
-class IParamProvider(object):
+class IParamProvider:
     def __call__(self, pyfunc, param_name):
         """
         :type pyfunc: rope.base.pyobjectsdef.PyFunction
@@ -8,7 +8,7 @@ class IParamProvider(object):
         raise NotImplementedError
 
 
-class IReturnProvider(object):
+class IReturnProvider:
     """
     :type resolve: rope.base.oi.type_hinting.resolvers.interfaces.IResolver
     """
@@ -23,7 +23,7 @@ class IReturnProvider(object):
         raise NotImplementedError
 
 
-class IAssignmentProvider(object):
+class IAssignmentProvider:
     """
     :type resolve: rope.base.oi.type_hinting.resolvers.interfaces.IResolver
     """