+++ rope/contrib/autoimport/sqlite.py
@@ -73,8 +73,8 @@ class AutoImport:
 
     connection: sqlite3.Connection
     underlined: bool
-    rope_project: Project
-    project: Package
+    project: Project
+    project_package: Package
 
     def __init__(self, project: Project, observe=True, underlined=False, memory=True):
         """Construct an AutoImport object.
@@ -90,19 +90,19 @@ class AutoImport:
         memory : bool
             if true, don't persist to disk
         """
-        self.rope_project = project
+        self.project = project
         project_package = get_package_tuple(
             pathlib.Path(project.root.real_path), project
         )
         assert project_package is not None
         assert project_package.path is not None
-        self.project = project_package
+        self.project_package = project_package
         self.underlined = underlined
         db_path: str
         if memory or project.ropefolder is None:
             db_path = ":memory:"
         else:
-            db_path = f"{project.ropefolder.path}/autoimport.db"
+            db_path = str(pathlib.Path(project.ropefolder.real_path) / "autoimport.db")
         self.connection = sqlite3.connect(db_path)
         self._setup_db()
         if observe:
@@ -195,7 +195,7 @@ class AutoImport:
         self, name: str, exact_match: bool = False
     ) -> Generator[SearchResult, None, None]:
         """
-        Search both names for avalible imports.
+        Search both names for available imports.
 
         Returns the import statement, import name, source, and type.
         """
@@ -217,7 +217,7 @@ class AutoImport:
         self, name: str, exact_match: bool = False
     ) -> Generator[SearchResult, None, None]:
         """
-        Search both modules for avalible imports.
+        Search both modules for available imports.
 
         Returns the import statement, import name, source, and type.
         """
@@ -284,7 +284,7 @@ class AutoImport:
         """
 
         if resources is None:
-            resources = self.rope_project.get_python_files()
+            resources = self.project.get_python_files()
         files = [Path(resource.real_path) for resource in resources]
         self._generate_cache(
             files=files, task_handle=task_handle, underlined=underlined
@@ -300,7 +300,7 @@ class AutoImport:
         """
         Generate global name cache for external modules listed in `modules`.
 
-        If no modules are provided, it will generate a cache for every module avalible.
+        If no modules are provided, it will generate a cache for every module available.
         This method searches in your sys.path and configured python folders.
         Do not use this for generating your own project's internal names,
         use generate_resource_cache for that instead.
@@ -392,9 +392,9 @@ class AutoImport:
         for module in modules:
             try:
                 module_name = module[0]
-                if module_name.startswith(f"{self.project.name}."):
+                if module_name.startswith(f"{self.project_package.name}."):
                     module_name = ".".join(module_name.split("."))
-                pymodule = self.rope_project.get_module(module_name)
+                pymodule = self.project.get_module(module_name)
                 if name in pymodule:
                     pyname = pymodule[name]
                     module, lineno = pyname.get_definition_location()
@@ -423,12 +423,12 @@ class AutoImport:
         if match is not None:
             code = code[: match.start()]
         try:
-            pymodule = libutils.get_string_module(self.rope_project, code)
+            pymodule = libutils.get_string_module(self.project, code)
         except exceptions.ModuleSyntaxError:
             return 1
         testmodname = "__rope_testmodule_rope"
         importinfo = importutils.NormalImport(((testmodname, None),))
-        module_imports = importutils.get_module_imports(self.rope_project, pymodule)
+        module_imports = importutils.get_module_imports(self.project, pymodule)
         module_imports.add_import(importinfo)
         code = module_imports.get_changed_source()
         offset = code.index(testmodname)
@@ -461,10 +461,12 @@ class AutoImport:
             self.connection.commit()
 
     def _get_python_folders(self) -> List[pathlib.Path]:
-        folders = self.rope_project.get_python_path_folders()
-        folder_paths = [
-            pathlib.Path(folder.path) for folder in folders if folder.path != "/usr/bin"
-        ]
+        def filter_folders(folder: pathlib.Path) -> bool:
+            return folder.is_dir() and folder.as_posix() != "/usr/bin"
+
+        folders = self.project.get_python_path_folders()
+        folder_paths = map(lambda folder: pathlib.Path(folder.real_path), folders)
+        folder_paths = filter(filter_folders, folder_paths)
         return list(OrderedDict.fromkeys(folder_paths))
 
     def _get_available_packages(self) -> List[Package]:
@@ -474,7 +476,7 @@ class AutoImport:
         ]
         for folder in self._get_python_folders():
             for package in folder.iterdir():
-                package_tuple = get_package_tuple(package, self.rope_project)
+                package_tuple = get_package_tuple(package, self.project)
                 if package_tuple is None:
                     continue
                 packages.append(package_tuple)
@@ -488,7 +490,7 @@ class AutoImport:
         existing: List[str] = list(
             chain(*self.connection.execute("select * from packages").fetchall())
         )
-        existing.append(self.project.name)
+        existing.append(self.project_package.name)
         return existing
 
     def _removed(self, resource):
@@ -521,7 +523,7 @@ class AutoImport:
             return Package(target_name, Source.BUILTIN, None, PackageType.BUILTIN)
         for folder in self._get_python_folders():
             for package in folder.iterdir():
-                package_tuple = get_package_tuple(package, self.rope_project)
+                package_tuple = get_package_tuple(package, self.project)
                 if package_tuple is None:
                     continue
                 name, source, package_path, package_type = package_tuple
@@ -531,13 +533,13 @@ class AutoImport:
         return None
 
     def _path_to_module(self, path: Path, underlined: bool = False) -> ModuleFile:
-        assert self.project.path
+        assert self.project_package.path
         underlined = underlined if underlined else self.underlined
         # The project doesn't need its name added to the path,
         # since the standard python file layout accounts for that
         # so we set add_package_name to False
         resource_modname: str = get_modname_from_path(
-            path, self.project.path, add_package_name=False
+            path, self.project_package.path, add_package_name=False
         )
         return ModuleFile(
             path,